src/requirements.txt
flask
mangum
asgiref
discord-interactions
boto3
pypdf
requests
python-dotenv
PyPDF2
openai
